% Generated by roxygen2 (4.0.1): do not edit by hand
\name{add_border}
\alias{add_border}
\title{Add a rectangular border around a plotting region}
\usage{
add_border(border = "black", ...)
}
\arguments{
\item{border}{Color for the rectangle's border}

\item{...}{Further arguments passed to \code{\link[graphics]{rect}}}
}
\value{
Returns the result of \code{par("usr")} - a vector of the
    form \code{c(x1, x2, y1, y2)} giving the extremes of the user
    coordinates of the plotting region.
}
\description{
Add a rectangular border around the region of a figure, to hide any small overlaps.
}
\examples{
x <- rnorm(100)
y <- 5*x + rnorm(100)
grayplot(x,y, pch=21, bg="slateblue")
abline(h=0, v=0, col="violetred")
add_border()
}
\author{
Karl W Broman, \email{kbroman@biostat.wisc.edu}
}

