% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_commas.R
\name{add_commas}
\alias{add_commas}
\title{Add commas to a large number}
\usage{
add_commas(numbers)
}
\arguments{
\item{numbers}{Vector of non-negative numbers (will be rounded to integers)}
}
\value{
Character string with numbers written like \code{"7,547,085"}.
}
\description{
Convert a number to a string, with commas every 3rd digit
}
\examples{
add_commas(c(231, 91310, 2123, 9911001020, 999723285))
}
