% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/align_vectors.R
\name{align_vectors}
\alias{align_vectors}
\title{Align two vectors}
\usage{
align_vectors(x, y, expand = TRUE)
}
\arguments{
\item{x}{A vector}

\item{y}{Another vector}

\item{expand}{If TRUE, expand each to the same length using NAs. If FALSE, remove elements not in common.}
}
\value{
A list with two components, \code{x} and \code{y}
}
\description{
Align two vectors using their names attributes, either expanding with NAs or reducing to the common values.
}
\examples{
x <- c(a=1, b=2, c=3, d=4)
y <- c(b=1, c=2, e=3, a=4)
align_vectors(x,y)
align_vectors(x,y,expand=FALSE)

}
