% Generated by roxygen2 (4.0.0): do not edit by hand
\name{attachwork}
\alias{attachwork}
\title{Attach a workspace}
\usage{
attachwork(i = "", dir = "Work", verbose = TRUE)
}
\arguments{
  \item{i}{An integer or character string.}

  \item{dir}{Directory name.}

  \item{verbose}{If true, print a message if the file can't
  be loaded.}
}
\value{
TRUE/FALSE according to whether the file exists (and so the function worked).
}
\description{
Attach a workspace for a directory name of a particular form.
}
\details{
This function attaches the workspace \code{paste(dir,i,"/.RData",sep="")}.
}
\examples{
\dontrun{for(i in 1:5) attachwork(i)}
}
\author{
Karl W Broman \email{kbroman@biostat.wisc.edu}
}
\seealso{
\code{\link{loadwork}},
  \code{\link{attachfile}}, \code{\link{loadfile}}
}
\keyword{IO}

