% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attrnames.R
\name{attrnames}
\alias{attrnames}
\title{Get names of attributes}
\usage{
attrnames(object)
}
\arguments{
\item{object}{Any object}
}
\value{
Vector of character strings with the names of the attributes.
}
\description{
Get the names of the attributes of an object
}
\details{
It just does \code{names(attributes(object))}.
}
\examples{
x <- matrix(1:100, ncol=5)
colnames(x) <- LETTERS[1:5]
attrnames(x)
}
\keyword{utilities}
