% Generated by roxygen2 (4.0.0): do not edit by hand
\name{brocolors}
\alias{brocolors}
\title{Vectors of colors for figures}
\usage{
brocolors(set = c("general", "bg", "bgpng", "CC", "f2", "sex", "main"))
}
\arguments{
  \item{set}{Character string indicating a set of colors.}
}
\value{
Vector of character strings representing the chosen set of colors, in RGB.
}
\description{
Creates different vectors of related colors that may be useful for figures.
}
\examples{
col <- brocolors("general")
plot(seq(along=col), pch=21, bg=col, cex=4)
}
\author{
Karl W Broman, \email{kbroman@biostat.wisc.edu}
}
\keyword{utilities}

