\name{broman-internal}
\alias{subtrap}
\alias{binary.v}

\title{Internal broman functions}

\description{
  Internal broman functions.
}

\usage{
subtrap(f, a, b, n=1, \dots)
binary.v(n)
}

\details{
  These are generally not to be called by the user.
}

\author{Karl W Broman \email{kbroman@biostat.wisc.edu} }

\keyword{internal}
