% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/broman-package.R
\docType{package}
\name{broman-package}
\alias{broman}
\alias{broman-package}
\title{broman: Karl Broman's R Code}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Miscellaneous R functions, including functions related to graphics (mostly for base graphics), permutation tests, running mean/median, and general utilities.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/kbroman/broman}
  \item Report bugs at \url{https://github.com/kbroman/broman/issues}
}

}
\author{
\strong{Maintainer}: Karl W Broman \email{broman@wisc.edu} (\href{https://orcid.org/0000-0002-4914-6671}{ORCID})

Authors:
\itemize{
  \item Karl W Broman \email{broman@wisc.edu} (\href{https://orcid.org/0000-0002-4914-6671}{ORCID})
}

Other contributors:
\itemize{
  \item Aimee Teo Broman (\href{https://orcid.org/0000-0003-3783-2807}{ORCID}) [contributor]
}

}
\keyword{internal}
