% Generated by roxygen2 (4.0.0): do not edit by hand
\name{bromanversion}
\alias{bromanversion}
\title{Installed version of R/broman}
\usage{
bromanversion()
}
\value{
A character string with the version number of the currently installed
  version of R/broman.
}
\description{
Print the version number of the currently installed version of R/broman.
}
\examples{
bromanversion()
}
\author{
Karl W Broman, \email{kbroman@biostat.wisc.edu}
}
\keyword{print}

