% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/excel_fig.R
\name{excel_fig}
\alias{excel_fig}
\title{Excel-style figure displaying contents of a matrix}
\usage{
excel_fig(mat, file, cellwidth = 100, cellheight = 20, pointsize = 14,
  border = "#CECECE", headcol = "#E9E9E9", headborder = "#969696",
  row_names = FALSE, col_names = TRUE, hilitcells, hilitcolor = "#F0DCDB")
}
\arguments{
\item{mat}{A matrix}

\item{file}{Optional file name (must have extension .svg, .png, .jpg, or .pdf)}

\item{cellwidth}{Width of each cell, in pixels}

\item{cellheight}{Height of each cell, in pixels}

\item{pointsize}{Pointsize for text (if \code{file} is provided)}

\item{border}{Color of border of cells for the body of the matrix}

\item{headcol}{Background color of cells on the top and left border}

\item{headborder}{Color of border of cells on the top and left border}

\item{row_names}{If TRUE, and row names are present, include them as a first column}

\item{col_names}{If TRUE, and column names are present, include them as a first row}

\item{hilitcells}{Character vector of cells to highlight, like \code{"A1"} or \code{"D4"}}

\item{hilitcolor}{Color to highlight cells, a vector of length 1 or the same length as \code{hilitcells}}
}
\description{
Turn a matrix of data into an SVG of how it might look in Excel
}
\examples{
df <- data.frame(id=    c(101,  102,  103),
                 sex=   c("M",  "F",  "M"),
                 weight=c(22.3, 15.8, 19.7),
                 stringsAsFactors=FALSE)
excel_fig(df, col_names=TRUE)
}
\keyword{hplot}

