% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_precision.R
\name{get_precision}
\alias{get_precision}
\title{Determine the precision of a number}
\usage{
get_precision(x, ...)
}
\arguments{
\item{x}{A numeric vector}

\item{...}{Ignore this}
}
\value{
A vector of integers, with the number of digits (to the last non-zero digit) past the decimal point.
}
\description{
Determine the precision of a number, as the number of digits past
the decimal point.
}
\details{
If the number is expressed in scientific notation, we take the number of digits
}
