% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert2hex.R
\name{hex2dec}
\alias{hex2dec}
\title{Convert from hex to decimal}
\usage{
hex2dec(h)
}
\arguments{
\item{h}{Vector of character strings with hexadecimal representation of integers
(values >= 2^31 converted to missing, \code{NA})}
}
\value{
The input converted from hexadecimal to decimal notation.
}
\description{
Convert a number from hexidecimal to decimal notation.
}
\examples{
hex2dec("14D")
hex2dec(0:30)

}
\seealso{
\code{\link[=dec2hex]{dec2hex()}}
}
\author{
Karl W Broman, \email{broman@wisc.edu}
}
\keyword{manip}
