% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/karl_theme.R
\name{karl_theme}
\alias{karl_theme}
\title{Karl's ggplot2 theme}
\usage{
karl_theme(...)
}
\arguments{
\item{...}{Passed to \code{\link[ggplot2]{theme}}}
}
\value{
An object as returned by \code{\link[ggplot2]{theme}}
}
\description{
Karl's ggplot2 theme: black border and no ticks
}
\examples{
library(ggplot2)
mtcars$cyl <- factor(mtcars$cyl)
ggplot(mtcars, aes(y=mpg, x=disp, color=cyl)) +
    geom_point() + karl_theme()
}
\seealso{
\code{\link[ggplot2]{theme}}
}

