% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kbdate.R
\name{kbdate}
\alias{kbdate}
\title{My little date facility}
\usage{
kbdate(format = c("dateonly", "standard"), date = Sys.time())
}
\arguments{
\item{format}{The format for the output}

\item{date}{The date/time to convert}
}
\value{
A character string representation of the date/time
}
\description{
Sys.Date as a string, in a few different formats
}
\examples{
kbdate()
kbdate("standard")

}
\seealso{
\code{\link[base:Sys.time]{base::Sys.time()}}, \code{\link[base:date]{base::date()}}
}
\keyword{utilities}
