% Generated by roxygen2 (4.0.0): do not edit by hand
\name{loadwork}
\alias{loadwork}
\title{Load a workspace}
\usage{
loadwork(i = "", dir = "Work", verbose = TRUE)
}
\arguments{
  \item{i}{An integer or character string.}

  \item{dir}{Directory name.}

  \item{verbose}{If true, print a message if the file can't
  be loaded.}
}
\value{
TRUE/FALSE according to whether the file exists (and so the function worked).
}
\description{
Load a workspace for a directory name of a particular form.
}
\details{
This function loads the workspace \code{paste(dir,i,"/.RData",sep="")}.
}
\examples{
\dontrun{for(i in 1:5) loadwork(i)}
}
\author{
Karl W Broman \email{kbroman@biostat.wisc.edu}
}
\seealso{
\code{\link{attachwork}},
  \code{\link{attachfile}},   \code{\link{loadfile}}
}
\keyword{IO}

