% Generated by roxygen2 (4.0.2.9000): do not edit by hand
% Please edit documentation in R/make.R
\name{make}
\alias{make}
\title{Run make within a package directory}
\usage{
make(pkg = ".", makefile = "Makefile", quiet = FALSE)
}
\arguments{
\item{pkg}{package description, can be path or package name.
See \code{\link[devtools]{as.package}} for more information}

\item{makefile}{File name of makefile}

\item{quiet}{If TRUE suppresses output from this function}
}
\value{
Exit value from \code{\link[base]{system}} with \code{intern=FALSE}
}
\description{
Run make within a package directory
}
\examples{
\dontrun{make() # run make within working directory
make("/path/to/mypackage") # run make within /path/to/mypackage
}
}
\seealso{
\code{\link[devtools]{load_all}}
}
\keyword{utilities}

