% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maxabs.R
\name{maxabs}
\alias{maxabs}
\title{maximum of absolute value}
\usage{
maxabs(x, na.rm = FALSE)
}
\arguments{
\item{x}{a numeric vector or array}

\item{na.rm}{a logical indicating whether missing values should be removed.}
}
\value{
The maximum of the absolute value of the input
}
\description{
Take the maximum of the absolute values of the input
}
\examples{
x <- c(5, -2, 8, -20, 2.3)
maxabs(x)
}
