% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pushbullet.R
\name{note}
\alias{note}
\title{Send a note to pushbullet}
\usage{
note(title, recipients = NULL, body = "")
}
\arguments{
\item{title}{The title of the note (could be the whole thing).}

\item{recipients}{A character or numeric vector indicating the
devices this post should go to. If NULL, the default device is
looked up from an optional setting, and if none has been set the
push is sent to all devices. (passed to
\code{\link[RPushbullet]{pbPost}}.)}

\item{body}{The body of the note (by default, empty)}
}
\description{
Even simpler interface for \code{\link[RPushbullet]{pbPost}}
to send a bit of text to pushbullet.
}
\examples{
\dontrun{note("Hello.")}
}
\keyword{utilities}
