% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/numbers.R
\docType{data}
\name{numbers}
\alias{numbers}
\alias{Numbers}
\title{Numbers spelled out in English}
\format{
A vector of character strings
}
\description{
The numbers 1-20 spelled out in English, for use in reports.
}
\details{
\itemize{
\item \code{numbers} - lower case
\item \code{Numbers} - Capitalized
}
}
\examples{
numbers[5]
Numbers[5]
}
\keyword{datasets}
