% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pastedot.R
\name{paste.}
\alias{paste.}
\title{paste with dot separator}
\usage{
paste.(...)
}
\arguments{
\item{...}{Passed to paste.}
}
\value{
A character string or vector of character strings.
}
\description{
Calls \code{\link[base:paste]{base::paste()}} with \code{sep="."}.
}
\details{
There's not much to this function.  It is just
\code{\link[base:paste]{base::paste()}} with \code{sep="."}, because I'm lazy.
}
\examples{
x <- 3
y <- 4
paste.(x, y)

}
\seealso{
\code{\link[base:paste]{base::paste()}},
\code{\link[base:paste]{base::paste0()}},
\code{\link[=paste00]{paste00()}},
\code{\link[=paste..]{paste..()}},
\code{\link[=paste0.]{paste0.()}},
\code{\link[=paste.0]{paste.0()}}
}
\keyword{character}
