% Generated by roxygen2 (4.0.0.99): do not edit by hand
\name{plot_crayons}
\alias{plot_crayons}
\title{Illustraion of crayon colors}
\usage{
plot_crayons(method2order = c("hclust", "mds", "svd"), colorspace = c("rgb",
  "lab", "hsv"), cex = 0.6)
}
\arguments{
\item{method2order}{method for ordering the colors in the plot
hierarchicical clustering (\code{"hclust"}), multi-dimensional scaling (\code{"mds"})}

\item{colorspace}{Color space (LAB or RGB) to use for ordering colors}

\item{cex}{character expansion for the text}
}
\value{
None
}
\description{
Creates a plot of the crayon colors in \code{\link{brocolors}}
}
\examples{
plot_crayons()
}
\author{
Karl W Broman, \email{kbroman@biostat.wisc.edu}
}
\references{
\url{http://en.wikipedia.org/wiki/List_of_Crayola_crayon_colors}
}
\seealso{
\code{\link{brocolors}}
}

