% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pushbullet.R
\name{pushbullet_devices}
\alias{pushbullet_devices}
\title{Grab info on Pushbullet devices.}
\usage{
pushbullet_devices()
}
\value{
data frame with nickname, model, and iden for active devices.
}
\description{
Get names and identifiers of Pushbullet devices.
}
\seealso{
\code{\link[=errors2pushbullet]{errors2pushbullet()}}, \code{\link[=done]{done()}}, \code{\link[=note]{note()}}
}
