% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/revgray.R
\name{revgray}
\alias{revgray}
\title{Create vector of colors from white to black}
\usage{
revgray(n = 256, ...)
}
\arguments{
\item{n}{Number of colors.}

\item{...}{Passed to \code{\link[grDevices:gray]{grDevices::gray()}}.}
}
\value{
Vector of colors, from white to black
}
\description{
Calls \code{\link[grDevices:gray]{grDevices::gray()}} then \code{\link[base:rev]{base::rev()}}
}
\details{
There's not much to this. It's just \code{gray((n:0)/n))}
}
\examples{
x <- matrix(rnorm(100), ncol=10)
image(x, col=revgray())

}
\seealso{
\code{\link[grDevices:gray]{grDevices::gray()}}
}
\keyword{color}
