% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setRNGparallel.R
\name{setRNGparallel}
\alias{setRNGparallel}
\alias{unsetRNGparallel}
\title{Set up random number generation for parallel calculations}
\usage{
setRNGparallel()

unsetRNGparallel()
}
\description{
Set random number generation to L'Ecuyer-CMRG, for use in parallel calculations.
}
\details{
I can never remember the command \code{RNGkind("L'Ecuyer-CMRG")}; this is a shortcut.
\code{unsetRNG4parallel} sets the random number generator back to the default type.
}
\examples{
RNGkind()
setRNGparallel()
RNGkind()
unsetRNGparallel()
RNGkind()
}
\keyword{utilities}
