% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spell_out.R
\name{spell_out}
\alias{spell_out}
\title{Spell out an integer}
\usage{
spell_out(number, capitalize = FALSE, max_value = 9)
}
\arguments{
\item{number}{A number that is to be spelled out (can be a vector).}

\item{capitalize}{If TRUE, capitalize the first letter.}

\item{max_value}{Maximum value to use (generally 9); if larger than this, use numerals.}
}
\value{
Character string (or vector of character strings) with numbers spelled out, or as numerals if large.
}
\description{
Spell out an integer as a word, for use in reports/papers.
}
\examples{
spell_out(9)
spell_out(9, cap=TRUE)
spell_out(9, max_value=5)
}
