% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pushbullet.R
\name{stop_sending_errors}
\alias{stop_sending_errors}
\title{Stop sending errors to pushbullet}
\usage{
stop_sending_errors()
}
\description{
Clear the \code{error} option, so that error notifications are no
longer sent to pushbullet.
}
\examples{
\dontrun{stop_sending_errors()}
}
\seealso{
\code{\link{errors2pushbullet}}
}
\keyword{utilities}

