% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_karl.R
\name{theme_karl}
\alias{theme_karl}
\alias{karl_theme}
\title{Karl's ggplot2 theme}
\usage{
theme_karl(base_size = 12, base_family = "", ...)

karl_theme(base_size = 12, base_family = "", ...)
}
\arguments{
\item{base_size}{Base font size}

\item{base_family}{Base font family}

\item{...}{Passed to \code{\link[ggplot2]{theme}}}
}
\value{
An object as returned by \code{\link[ggplot2]{theme}}
}
\description{
Karl's ggplot2 theme: black border and no ticks
}
\examples{
library(ggplot2)
mtcars$cyl <- factor(mtcars$cyl)
ggplot(mtcars, aes(y=mpg, x=disp, color=cyl)) +
    geom_point() + theme_karl()

}
\seealso{
\code{\link[ggplot2]{theme}}
}
