% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time_axis.R
\name{time_axis}
\alias{time_axis}
\title{Set up a time-based axis}
\usage{
time_axis(times, n = 8)
}
\arguments{
\item{times}{A vector of date/times that will be plotted}

\item{n}{Number of values to use in axis}
}
\value{
A data frame with the numeric values to plot plus labels to use.
}
\description{
Set up a time-based axis for base graphics
}
