% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viridis.R
\name{viridis}
\alias{viridis}
\title{viridis color scheme}
\usage{
viridis(n = 256, alpha = 1, begin = 0, end = 1)
}
\arguments{
\item{n}{Number of colors}

\item{alpha}{Transparency}

\item{begin}{Number for starting color}

\item{end}{Number for ending color}
}
\value{
Vector of color codes, of length \code{n}.
}
\description{
Vector of colors following the viridis color scheme.
}
\examples{
viridis(20)
}
\references{
Code modified from \url{https://github.com/sjmgarnier/viridis} by Simon Garnier
}
\keyword{htest}
