% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/align_stuff.R
\name{align_vectors}
\alias{align_vectors}
\title{Align two vectors}
\usage{
align_vectors(x, y, expand = TRUE)
}
\arguments{
\item{x}{A vector}

\item{y}{Another vector}

\item{expand}{If TRUE, expand each to the same length using NAs. If FALSE, remove elements not in common.}
}
\value{
A list with two components, \code{x} and \code{y}
}
\description{
Align two vectors using their names attributes, either expanding with NAs or reducing to the common values.
}
