% Generated by roxygen2 (4.0.1): do not edit by hand
\name{attachfile}
\alias{attachfile}
\title{Attach a workspace}
\usage{
attachfile(i, stem = "perm", end = ".RData", fixdig = TRUE,
  max_digits = 5, verbose = TRUE)
}
\arguments{
\item{i}{An integer or character string.}

\item{stem}{Initial part of name.}

\item{end}{Last part of name.}

\item{fixdig}{If TRUE, pad \code{i} with initial 0's until file name exists.}

\item{max_digits}{If \code{figdig=TRUE}, maximum number of 0's to pad}

\item{verbose}{If true, print a message if the file can't be loaded.}
}
\value{
TRUE/FALSE according to whether the file exists (and so the function worked).
}
\description{
Attach a workspace for a directory name of a particular form.
}
\details{
This function attaches the workspace \code{paste(stem,i,end,sep="")},
  possibly adding 0's before i.
}
\examples{
\dontrun{for(i in 1:5) attachfile(i)}
}
\seealso{
\code{\link{loadwork}},
  \code{\link{attachwork}}, \code{\link{loadfile}}
}
\keyword{IO}

