\name{bromanversion}
\alias{bromanversion}

\title{Installed version of R/broman}

\description{
  Print the version number of the currently installed version of R/broman.
}

\usage{bromanversion()}

%\arguments{
%}

\value{
  A character string with the version number of the currently installed
  version of R/broman.
}

\author{Karl W Broman, \email{kbroman@biostat.wisc.edu} }

\examples{
  bromanversion()
}


\keyword{print}
