% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{bromanversion}
\alias{bromanversion}
\title{Installed version of R/broman}
\usage{
bromanversion()
}
\value{
A character string with the version number of the currently installed
version of R/broman.
}
\description{
Print the version number of the currently installed version of R/broman.
}
\examples{
bromanversion()

}
\keyword{print}
