\name{convert2hex}
\alias{convert2hex}

\title{Convert to hex}

\description{
  Convert a number to hexidecimal notation.
}

\usage{
convert2hex(d)
}

\arguments{
 \item{d}{A number.}
}

\value{
  A character string; the input in hex.
}

\details{
  Nothing important to say here.
}

\author{Karl W Broman, \email{kbroman@biostat.wisc.edu} }

\keyword{manip}
