% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert2hex.R
\name{convert2hex}
\alias{convert2hex}
\alias{dec2hex}
\title{Convert decimal to hex}
\usage{
convert2hex(d)
}
\arguments{
\item{d}{A number.}
}
\value{
A character string; the input in hex.
}
\description{
Convert a number to hexidecimal notation.
}
\details{
Nothing important to say here.
}
\examples{
convert2hex(333)
dec2hex(333)
dec2hex(333) == "14D"
dec2hex(0:30)

}
\seealso{
\code{\link[=hex2dec]{hex2dec()}}
}
\keyword{manip}
