% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brocolors.R
\name{crayons}
\alias{crayons}
\title{Crayon colors}
\usage{
crayons(color_names = NULL, ...)
}
\arguments{
\item{color_names}{Optional vector of color names; can be partial matches.}

\item{...}{Additional optional color names}
}
\value{
Vector of named RGB colors
}
\description{
Vector of colors corresponding to Crayola crayons
}
\references{
\url{http://en.wikipedia.org/wiki/List_of_Crayola_crayon_colors}
}
\seealso{
\code{\link[=plot_crayons]{plot_crayons()}}, \code{\link[=brocolors]{brocolors()}}
}
\keyword{utilities}
