% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/pushbullet.R
\name{done}
\alias{done}
\title{Send a short message via RPushbullet.}
\usage{
done(message = "R is done", recipients)
}
\arguments{
\item{message}{A character string with a message.
(passed to \code{\link[RPushbullet]{pbPost}}.)}

\item{recipients}{A character or numeric vector indicating the
devices this post should go to. If missing, the default device is
looked up from an optional setting, and if none has been set the
push is sent to all devices. (passed to
\code{\link[RPushbullet]{pbPost}}.)}
}
\description{
Send a short message via RPushbullet, to be used to indicate that
some R job is complete.
#'
}
\examples{
\dontrun{done("Your R job is complete.")}
}
\keyword{utilities}

