% Generated by roxygen2 (4.0.1): do not edit by hand
\name{errors2pushbullet}
\alias{errors2pushbullet}
\title{Send further errors to pushbullet}
\usage{
errors2pushbullet()
}
\description{
Set options to use RPushbullet to use pushbullet to push
notifications of any error messages.
}
\examples{
\dontrun{errors2pushbullet()}
}
\keyword{utilities}

