% Generated by roxygen2 (4.0.2.9000): do not edit by hand
% Please edit documentation in R/errors2pushbullet.R
\name{errors2pushbullet}
\alias{errors2pushbullet}
\title{Send further errors to pushbullet}
\usage{
errors2pushbullet(deviceind = 1)
}
\arguments{
\item{deviceind}{Numeric index for device to which errors will be
pushed; passed to \code{\link[RPushbullet]{pbPost}}.}
}
\description{
Set options to use RPushbullet to use pushbullet to push
notifications of any error messages.
}
\examples{
\dontrun{errors2pushbullet()}
}
\keyword{utilities}

