% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pushbullet.R
\name{errors2pushbullet}
\alias{errors2pushbullet}
\title{Send further errors to pushbullet}
\usage{
errors2pushbullet(recipients = NULL)
}
\arguments{
\item{recipients}{A character or numeric vector indicating the
devices this post should go to. If NULL, the default device is
looked up from an optional setting, and if none has been set the
push is sent to all devices. (passed to
\code{\link[RPushbullet]{pbPost}}.)}
}
\description{
Set options to use RPushbullet to use pushbullet to push
notifications of any error messages.
}
\examples{
\dontrun{errors2pushbullet()}
}
\seealso{
\code{\link{stop_sending_errors}}
}
\keyword{utilities}
