% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/get.R
\name{get0}
\alias{get.}
\alias{get0}
\title{get with paste}
\usage{
get0(...)
}
\arguments{
\item{...}{Vector of character strings.}
}
\value{
The object found.
}
\description{
Calls \code{\link[base]{paste}} then \code{\link[base]{get}}
}
\details{
There's not much to these functions.  \code{get0} uses \code{\link[base]{paste0}}
  and so combines its arguments with \code{sep=""}.   \code{get.} uses
  \code{\link[base]{paste}} with \code{sep="."}.
}
\examples{
xy <- 3
x.y <- 18.3
get0("x", "y")
get.("x", "y")
}
\seealso{
\code{\link[base]{get}},
  \code{\link[base]{paste}},
  \code{\link[base]{paste0}}
}
\keyword{data}

