% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/help.R
\name{h}
\alias{h}
\title{View html version of help file}
\usage{
h(...)
}
\arguments{
\item{...}{Help topics.}
}
\value{
No return value.
}
\description{
View the html version of a help file while running R via ESS within emacs.
}
\details{
This just calls the function \code{\link[utils]{help}} using the
  argument \code{htmlhelp=TRUE}.
}
\examples{
h(read.cross)

}
\seealso{
\code{\link[utils]{help}}, \code{\link[utils]{help.start}}
}
\keyword{documentation}

