% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/convert2hex.R
\name{hex2dec}
\alias{hex2dec}
\title{Convert from hex to decimal}
\usage{
hex2dec(h)
}
\arguments{
\item{h}{#' Character string with hexadecimal representation of a number}
}
\value{
The input converted from hexadecimal to decimal notation.
}
\description{
Convert a number from hexidecimal to decimal notation.
}
\details{
Nothing important to say here.
}
\examples{
hex2dec("14D")
hex2dec("14D") == 333
hex2dec(0:30)
}
\author{
Karl W Broman, \email{kbroman@biostat.wisc.edu}
}
\seealso{
\code{\link{dec2hex}}
}
\keyword{manip}

