% Generated by roxygen2 (4.0.0): do not edit by hand
\name{kbdate}
\alias{kbdate}
\title{My only little date facility}
\usage{
kbdate(format = c("dateonly", "standard"), date = Sys.time())
}
\arguments{
\item{format}{The format for the output}

\item{date}{The date/time to convert}
}
\value{
A character string representation of the date/time
}
\description{
Sys.Date as a string, in a few different formats
}
\examples{
kbdate()
kbdate("standard")
}
\author{
Karl W Broman \email{kbroman@biostat.wisc.edu}
}
\seealso{
\code{\link[base]{Sys.time}}, \code{\link[base]{date}}
}
\keyword{utilities}

