% Generated by roxygen2 (4.0.1): do not edit by hand
\name{loadfile}
\alias{loadfile}
\title{Load a workspace}
\usage{
loadfile(i, stem = "perm", end = ".RData", fixdig = TRUE, maxdig = 5,
  verbose = TRUE)
}
\arguments{
\item{i}{An integer or character string.}

\item{stem}{Initial part of name.}

\item{end}{Last part of name.}

\item{fixdig}{If TRUE, pad \code{i} with 0's until file exists (up to maxdig)}

\item{maxdig}{Maximum number of 0's to add}

\item{verbose}{If true, print a message if the file can't be loaded.}
}
\value{
TRUE/FALSE according to whether the file was loaded.
}
\description{
Load a workspace for a directory name of a particular form.
}
\details{
This function loads the workspace \code{paste(stem,i,end,sep="")},
  possibly adding 0's before i
}
\examples{
\dontrun{for(i in 1:5) loadfile(i)}
}
\author{
Karl W Broman \email{kbroman@biostat.wisc.edu}
}
\seealso{
\code{\link{attachfile}}
}
\keyword{IO}

