% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make.R
\name{make}
\alias{make}
\title{Run make within a package directory}
\usage{
make(pkg = ".", makefile = "Makefile", target = "", quiet = FALSE)
}
\arguments{
\item{pkg}{Path to directory containing the GNU Make file, or an
Rpackage description, which can be a path or a package name. (See
\code{\link[devtools]{as.package}} for more information.)}

\item{makefile}{File name of makefile.}

\item{target}{Optional character string specifying the target.}

\item{quiet}{If TRUE suppresses output from this function.}
}
\value{
Exit value from \code{\link[base]{system}} with \code{intern=FALSE}
}
\description{
Run make within a package directory
}
\examples{
\dontrun{make() # run make within working directory
make("/path/to/mypackage") # run make within /path/to/mypackage
}

}
\seealso{
\code{\link[devtools]{load_all}}
}
\keyword{utilities}
