% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myround.R
\name{myround}
\alias{myround}
\title{Round a number, preserving extra 0's}
\usage{
myround(x, digits = 1)
}
\arguments{
\item{x}{Number to round.}

\item{digits}{Number of digits past the decimal point to keep.}
}
\value{
A vector of character strings.
}
\description{
Round a number, preserving extra 0's.
}
\details{
Uses \code{\link[base]{sprintf}} to round a number, keeping extra 0's.
}
\examples{
myround(51.01, 3)
myround(0.199, 2)

}
\seealso{
\code{\link[base]{round}}, \code{\link[base]{sprintf}}
}
\keyword{utilities}

