% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/openfile.R
\name{openfile}
\alias{openfile}
\title{Open a file}
\usage{
openfile(file)
}
\arguments{
\item{file}{File name (character string)}
}
\value{
None.
}
\description{
Open a file using \code{\link[base]{system}} and \code{open}.
}
\details{
I'd thought that to open a file you'd use \code{open} in
MacOS and \code{start} in Windows, but \code{system("start
myfile.pdf")} doesn't work in Windows, but rather
\code{system("open myfile.pdf")} does, so here we're just using
}
\examples{
\dontrun{openfile("myplot.pdf")}
}
\keyword{IO}

