% Generated by roxygen2 (4.0.1): do not edit by hand
\name{paste.}
\alias{paste.}
\title{paste with dot separator}
\usage{
paste.(...)
}
\arguments{
\item{...}{Passed to paste.}
}
\value{
A character string or vector of character strings.
}
\description{
Calls \code{\link[base]{paste}} with \code{sep="."}.
}
\details{
There's not much to this function.  It just is
  \code{\link[base]{paste}} with \code{sep=""}, 'cause I'm lazy.
}
\examples{
x <- 3
y <- 4
paste.(x, y)
}
\seealso{
\code{\link[base]{paste}},
  \code{\link[base]{paste0}},
  \code{\link{paste00}},
  \code{\link{paste..}},
  \code{\link{paste0.}},
  \code{\link{paste.0}}
}
\keyword{character}

