% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pastedot.R
\name{paste00}
\alias{paste00}
\alias{paste0.}
\alias{paste.0}
\alias{paste..}
\title{paste with null or dot as separator and with collapse}
\usage{
paste00(...)
}
\arguments{
\item{...}{Passed to paste.}
}
\value{
A character string or vector of character strings.
}
\description{
Call \code{\link[base:paste]{base::paste()}} with \code{sep="."} or \code{sep=""}
and \code{collapse=""} or \code{collapse="."}.
}
\details{
There's not much to these functions.
\code{paste00(...)} is like \code{paste(..., sep="", collapse="")}
\code{paste..(...)} is like \code{paste(..., sep=".", collapse=".")}
\code{paste0.(...)} is like \code{paste(..., sep="", collapse=".")}
\code{paste.0(...)} is like \code{paste(..., sep=".", collapse="")}
}
\examples{
x <- c(3, 4)
y <- c(5, 6)
paste00(x, y)
paste..(x, y)
paste0.(x, y)
paste.0(x, y)

}
\seealso{
\code{\link[base:paste]{base::paste()}},
\code{\link[base:paste0]{base::paste0()}},
\code{\link[=paste.]{paste.()}}
}
\keyword{character}
