% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/errors2pushbullet.R
\name{pushbullet_devices}
\alias{pushbullet_devices}
\title{Grab info on Pushbullet devices.}
\usage{
pushbullet_devices(json = FALSE)
}
\value{
data frame with nickname, model, and iden for active devices.
}
\description{
Get names and identifiers of Pushbullet devices.
}

